+++ src/test/java/io/jenkins/plugins/analysis/core/testutil/FormValidationAssert.java
@@ -2,6 +2,7 @@ package io.jenkins.plugins.analysis.core.testutil;
 
 import java.util.Objects;
 
+import org.apache.commons.text.StringEscapeUtils;
 import org.assertj.core.api.AbstractAssert;
 
 import hudson.util.FormValidation;
@@ -86,8 +87,10 @@ public class FormValidationAssert extends AbstractAssert<FormValidationAssert, F
     public FormValidationAssert hasMessage(final String expectedMessage) {
         isNotNull();
 
-        if (!Objects.equals(actual.getMessage(), expectedMessage)) {
-            failWithMessage(EXPECTED_BUT_WAS_MESSAGE, "message", actual, expectedMessage, actual.getMessage());
+        String actualMessage = StringEscapeUtils.unescapeHtml4(actual.getMessage());
+        if (!Objects.equals(actualMessage, expectedMessage)) {
+            failWithMessage(EXPECTED_BUT_WAS_MESSAGE, "message", StringEscapeUtils.unescapeHtml4(actual.toString()), expectedMessage,
+                    actualMessage);
         }
 
         return this;